+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Object.path(obj, path) now also accepts an array or multiple arguments
 * Add Array.range(start, stop, step) method, which behaves like Python's range()
 * Add Array#closest(goal) method to find the closest value to the given goal
+* Added String#before(needle, first) and String#after(needle, first)
 
 ## 0.1.1 (2014-06-23)